Add a git-show command
